requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
jieba==0.42.1
wordcloud==1.9.2
matplotlib==3.7.2
Pillow==10.0.0
fake-useragent==1.4.0
numpy==1.24.3
